requirements.txt
fastapi
uvicorn
cryptography
python-dotenv

#uvicorn main:app --reload